Refactor [] Refactor code
