+++ Switcher/Assets.xcassets/icon.imageset/Image.png
